requirements.txt
coveralls==1.6.0
Flask==1.0.2
flask_sqlalchemy==2.3.2
flask_script==2.0.6
pip-licenses==1.10.0
PostgreSQL-Audit==0.10.0
psycopg2-binary==2.7.6.1
pytest==4.0.0
pytest-cov==2.6.1
pytest-flask==0.14.0
pytest-mock==1.10.0
sqlalchemy==1.2.11
SQLAlchemy==1.2.11
twine==1.13.0
